requirements.txt
discord.py==1.7.3
google_api_python_client==2.33.0
gTTS==2.2.3
imageio==2.13.3
moviepy==1.0.3
mutagen==1.45.1
Pillow==8.4.0
playsound==1.3.0
protobuf==3.19.1
PyDrive==1.3.1
pydub==0.25.1
requests==2.26.0
SQLAlchemy==1.4.28
httplib2<0.16.0